Lab08/shop_back/requirements.txt
Django==5.0.5
djangorestframework==3.14.0 